backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.1
pymongo==4.6.1
pycryptodome==3.20.0
web3==7.0.0
requests==2.31.0
gunicorn==21.2.0
gevent==23.9.1
gevent-websocket==0.10.1